requirements.txt
fastapi==0.111.1
pyjwt==2.8.0
motor==3.5.1
numpy==1.24.3
pandas==2.0.3
passlib==1.7.4
pydantic==2.8.2
pymongo==4.7.3
python_jose==3.3.0
scikit_learn==1.3.2
sentence_transformers==2.2.2
soynlp==0.0.493
starlette==0.37.2
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
uvicorn==0.18.3
bcrypt==4.1.3